requirements.txt
amqp==2.5.1
beautifulsoup4==4.8.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
django-redis==4.10.0
djangorestframework==3.10.2
idna==2.8
importlib-metadata==0.20
kombu==4.6.4
more-itertools==7.2.0
Pillow==6.1.0
psycopg2==2.8.3
pytz==2019.2
redis==3.3.8
requests==2.22.0
soupsieve==1.9.3
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
zipp==0.6.0
